+++ cognite/client/data_classes/data_modeling/data_types.py
@@ -1,21 +1,23 @@
 from __future__ import annotations
 
-import inspect
 import logging
-from abc import ABC, abstractmethod
+from abc import ABC
 from dataclasses import asdict, dataclass
-from typing import Any, ClassVar, cast
+from typing import TYPE_CHECKING, Any, ClassVar
 
-from typing_extensions import Self
+from typing_extensions import Self, TypeAlias
 
+from cognite.client.data_classes._base import CogniteObject, UnknownCogniteObject
 from cognite.client.data_classes.data_modeling.ids import ContainerId
 from cognite.client.utils._auxiliary import rename_and_exclude_keys
-from cognite.client.utils._text import convert_all_keys_recursive, convert_all_keys_to_snake_case
+from cognite.client.utils._text import convert_all_keys_recursive
+
+if TYPE_CHECKING:
+    from cognite.client import CogniteClient
 
 logger = logging.getLogger(__name__)
 
-_PROPERTY_ALIAS = {"list": "isList"}
-_PROPERTY_ALIAS_INV = {"isList": "list", "is_list": "list"}
+_PROPERTY_ALIAS = {"isList": "list", "is_list": "list"}
 
 
 @dataclass(frozen=True)
@@ -46,8 +48,9 @@ class DirectRelationReference:
 
 
 @dataclass
-class PropertyType(ABC):
+class PropertyType(CogniteObject, ABC):
     _type: ClassVar[str]
+    is_list: bool = False
 
     def dump(self, camel_case: bool = True) -> dict[str, Any]:
         output = asdict(self)
@@ -55,71 +58,86 @@ class PropertyType(ABC):
         for key in list(output.keys()):
             if output[key] is None:
                 output.pop(key)
-        output = rename_and_exclude_keys(output, aliases=_PROPERTY_ALIAS_INV)
+        output = rename_and_exclude_keys(output, aliases=_PROPERTY_ALIAS)
         return convert_all_keys_recursive(output, camel_case)
 
+    @staticmethod
+    def __load_unit_ref(data: dict) -> UnitReference | None:
+        unit: UnitReference | None = None
+        if (unit_raw := data.get("unit")) and isinstance(unit_raw, dict):
+            unit = UnitReference.load(unit_raw)
+        return unit
+
     @classmethod
-    def load(cls, data: dict) -> Self:
-        if "type" not in data:
-            raise ValueError("Property types are required to have a type")
-        type_ = data["type"]
-        data = convert_all_keys_to_snake_case(rename_and_exclude_keys(data, aliases=_PROPERTY_ALIAS, exclude={"type"}))
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> Self:
+        type_ = resource["type"]
+        obj: Any
+        if type_ == "text":
+            obj = Text(is_list=resource["list"], collation=resource["collation"])
+        elif type_ == "boolean":
+            obj = Boolean(is_list=resource["list"])
+        elif type_ == "float32":
+            obj = Float32(is_list=resource["list"], unit=cls.__load_unit_ref(resource))
+        elif type_ == "float64":
+            obj = Float64(is_list=resource["list"], unit=cls.__load_unit_ref(resource))
+        elif type_ == "int32":
+            obj = Int32(is_list=resource["list"], unit=cls.__load_unit_ref(resource))
+        elif type_ == "int64":
+            obj = Int64(is_list=resource["list"], unit=cls.__load_unit_ref(resource))
+        elif type_ == "timestamp":
+            obj = Timestamp(is_list=resource["list"])
+        elif type_ == "date":
+            obj = Date(is_list=resource["list"])
+        elif type_ == "json":
+            obj = Json(is_list=resource["list"])
+        elif type_ == "timeseries":
+            obj = TimeSeriesReference(is_list=resource["list"])
+        elif type_ == "file":
+            obj = FileReference(is_list=resource["list"])
+        elif type_ == "sequence":
+            obj = SequenceReference(is_list=resource["list"])
+        elif type_ == "direct":
+            obj = DirectRelation(
+                container=ContainerId.load(container) if (container := resource.get("container")) else None,
+                is_list=resource["list"],
+            )
+        else:
+            logger.warning(f"Unknown property type: {type_}")
+            obj = UnknownCogniteObject(resource)
+        return obj
 
-        if type_cls := _TYPE_LOOKUP.get(type_):
-            if issubclass(type_cls, LoadablePropertyType):
-                return cast(Self, type_cls.load(data))
-            try:
-                return cast(Self, type_cls(**data))
-            except TypeError:
-                not_supported = set(data).difference(inspect.signature(type_cls).parameters) - {"type"}
-                logger.warning(
-                    f"For '{type_cls.__name__}', the following properties are not yet supported in the SDK (ignored): "
-                    f"{not_supported}. Try updating to the latest SDK version, or create an issue on Github!"
-                )
-                return cast(Self, type_cls(**rename_and_exclude_keys(data, exclude=not_supported)))
 
-        raise ValueError(f"Invalid type {type_}.")
+# Kept around for backwards compatibility
+UnknownPropertyType: TypeAlias = UnknownCogniteObject
 
 
 @dataclass
-class LoadablePropertyType(ABC):
-    @classmethod
-    @abstractmethod
-    def load(cls, data: dict) -> Self: ...
-
-
-@dataclass
-class ListablePropertyType(PropertyType, ABC):
-    is_list: bool = False
-
-
-@dataclass
-class Text(ListablePropertyType):
+class Text(PropertyType):
     _type = "text"
     collation: str = "ucs_basic"
 
 
 @dataclass
-class Primitive(ListablePropertyType, ABC): ...
+class Primitive(PropertyType, ABC): ...
 
 
 @dataclass
-class Boolean(ListablePropertyType):
+class Boolean(PropertyType):
     _type = "boolean"
 
 
 @dataclass
-class Timestamp(ListablePropertyType):
+class Timestamp(PropertyType):
     _type = "timestamp"
 
 
 @dataclass
-class Date(ListablePropertyType):
+class Date(PropertyType):
     _type = "date"
 
 
 @dataclass
-class Json(ListablePropertyType):
+class Json(PropertyType):
     _type = "json"
 
 
@@ -155,19 +173,8 @@ class UnitSystemReference:
 
 
 @dataclass
-class ListablePropertyTypeWithUnit(ListablePropertyType, LoadablePropertyType, ABC):
+class PropertyTypeWithUnit(PropertyType, ABC):
     unit: UnitReference | None = None
-
-    @classmethod
-    def load(cls, data: dict) -> Self:
-        data = convert_all_keys_to_snake_case(rename_and_exclude_keys(data, aliases=_PROPERTY_ALIAS, exclude={"type"}))
-        unit: UnitReference | None = None
-        if (unit_raw := data.get("unit")) and isinstance(unit_raw, dict):
-            unit = UnitReference.load(unit_raw)
-        return cls(
-            is_list=data["is_list"],
-            unit=unit,
-        )
 
     def dump(self, camel_case: bool = True) -> dict[str, Any]:
         output = super().dump(camel_case)
@@ -177,27 +184,27 @@ class ListablePropertyTypeWithUnit(ListablePropertyType, LoadablePropertyType, A
 
 
 @dataclass
-class Float32(ListablePropertyTypeWithUnit):
+class Float32(PropertyTypeWithUnit):
     _type = "float32"
 
 
 @dataclass
-class Float64(ListablePropertyTypeWithUnit):
+class Float64(PropertyTypeWithUnit):
     _type = "float64"
 
 
 @dataclass
-class Int32(ListablePropertyType):
+class Int32(PropertyTypeWithUnit):
     _type = "int32"
 
 
 @dataclass
-class Int64(ListablePropertyType):
+class Int64(PropertyTypeWithUnit):
     _type = "int64"
 
 
 @dataclass
-class CDFExternalIdReference(ListablePropertyType, ABC): ...
+class CDFExternalIdReference(PropertyType, ABC): ...
 
 
 @dataclass
@@ -216,7 +223,7 @@ class SequenceReference(CDFExternalIdReference):
 
 
 @dataclass
-class DirectRelation(PropertyType, LoadablePropertyType):
+class DirectRelation(PropertyType):
     _type = "direct"
     container: ContainerId | None = None
 
@@ -228,24 +235,3 @@ class DirectRelation(PropertyType, LoadablePropertyType):
             elif output["container"] is None:
                 output.pop("container")
         return output
-
-    @classmethod
-    def load(cls, data: dict) -> Self:
-        return cls(container=ContainerId.load(container) if (container := data.get("container")) else None)
-
-
-_TYPE_LOOKUP: dict[str, type[PropertyType]] = {
-    "text": Text,
-    "boolean": Boolean,
-    "float32": Float32,
-    "float64": Float64,
-    "int32": Int32,
-    "int64": Int64,
-    "timestamp": Timestamp,
-    "date": Date,
-    "json": Json,
-    "timeseries": TimeSeriesReference,
-    "file": FileReference,
-    "sequence": SequenceReference,
-    "direct": DirectRelation,
-}