+++ cognite/client/data_classes/data_modeling/instances.py
@@ -77,7 +77,7 @@ PropertyIdentifier: TypeAlias = str
 
 
 @dataclass
-class NodeOrEdgeData:
+class NodeOrEdgeData(CogniteObject):
     """This represents the data values of a node or edge.
 
     Args:
@@ -89,21 +89,21 @@ class NodeOrEdgeData:
     properties: Mapping[str, PropertyValue]
 
     @classmethod
-    def load(cls, data: dict) -> NodeOrEdgeData:
+    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> Self:
         try:
-            source_type = data["source"]["type"]
+            source_type = resource["source"]["type"]
         except KeyError as e:
             raise ValueError("source must be a dict with a type key") from e
         source: ContainerId | ViewId
         if source_type == "container":
-            source = ContainerId.load(data["source"])
+            source = ContainerId.load(resource["source"])
         elif source_type == "view":
-            source = ViewId.load(data["source"])
+            source = ViewId.load(resource["source"])
         else:
             raise ValueError(f"source type must be container or view, but was {source_type}")
         return cls(
             source=source,
-            properties=data["properties"],
+            properties=resource["properties"],
         )
 
     def dump(self, camel_case: bool = True) -> dict:
@@ -996,10 +996,6 @@ class InstancesResult:
 
     nodes: NodeList
     edges: EdgeList
-
-    @classmethod
-    def load(cls, data: str | dict) -> InstancesResult:
-        raise NotImplementedError
 
 
 @dataclass