+++ tests/tests_unit/test_data_classes/test_datapoints.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import math
+
 import pytest
 
 from cognite.client.data_classes import DatapointsArray
@@ -19,9 +21,9 @@ def factory_method_from_array_data():
     yield pytest.param([arr1, arr2], expected, id="Construct from two arrays")
 
 
+@pytest.mark.dsl
 class TestDatapointsArray:
     @staticmethod
-    @pytest.mark.dsl
     @pytest.mark.parametrize("arrays, expected_array", list(factory_method_from_array_data()))
     def test_factory_method_from_array(arrays: list[DatapointsArray], expected_array: DatapointsArray):
         import numpy as np
@@ -30,3 +32,18 @@ class TestDatapointsArray:
 
         np.testing.assert_allclose(actual_array.average, expected_array.average)
         np.testing.assert_equal(actual_array.timestamp, expected_array.timestamp)
+
+    def test_dump_converts_missing_values_to_none(self):
+        # Easy to forget that we can have bad data (missing) without any status codes on the object
+        import numpy as np
+
+        params = dict(
+            id=123,
+            timestamp=np.array([1, 2, 3], dtype=np.int64),
+            value=np.array([-1, None, 2.5], dtype=np.float64),
+        )
+        dps1 = DatapointsArray(**params).dump()
+        dps2 = DatapointsArray(**params, null_timestamps={2}).dump()
+        assert dps1 != dps2
+        assert math.isnan(dps1["datapoints"][1]["value"])
+        assert dps2["datapoints"][1]["value"] is None